+++ spotbugs/src/main/java/edu/umd/cs/findbugs/AbstractBugReporter.java
@@ -26,6 +26,7 @@ import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 import javax.annotation.CheckForNull;
@@ -132,6 +133,8 @@ public abstract class AbstractBugReporter implements BugReporter {
 
     private final ProjectStats projectStats;
 
+    private PriorityAdjuster priorityAdjuster;
+
     protected AbstractBugReporter() {
         super();
         verbosityLevel = NORMAL;
@@ -142,6 +145,8 @@ public abstract class AbstractBugReporter implements BugReporter {
         // bug 2815983: no bugs are reported anymore
         // there is no info which value should be default, so using the max
         rankThreshold = BugRanker.VISIBLE_RANK_MAX;
+        // using by default empty settings
+        priorityAdjuster = new PriorityAdjuster(Map.of());
     }
 
     @Override
@@ -163,6 +168,17 @@ public abstract class AbstractBugReporter implements BugReporter {
         this.relaxedSet = true;
     }
 
+    /**
+     * Allows to adjust priorities when bugs are reported. Bugs imported from XML
+     * (see {@link #reportBugsFromXml(InputStream, Project)} are not affected.
+     *
+     * @param priorityAdjuster the priority adjuster
+     */
+    @Override
+    public void setPriorityAdjuster(PriorityAdjuster priorityAdjuster) {
+        this.priorityAdjuster = priorityAdjuster;
+    }
+
     protected boolean isRelaxed() {
         if (!relaxedSet) {
             if (FindBugsAnalysisFeatures.isRelaxedMode()) {
@@ -181,6 +197,7 @@ public abstract class AbstractBugReporter implements BugReporter {
             doReportBug(bugInstance);
             return;
         }
+        bugInstance = priorityAdjuster != null ? priorityAdjuster.adjustPriority(bugInstance) : bugInstance;
         if (priorityThreshold == 0) {
             throw new IllegalStateException("Priority threshold not set");
         }
@@ -421,4 +438,9 @@ public abstract class AbstractBugReporter implements BugReporter {
      *            the name of the class
      */
     public abstract void reportMissingClass(String string);
+
+    @Override
+    public PriorityAdjuster getPriorityAdjuster() {
+        return priorityAdjuster;
+    }
 }