+++ spotbugs/src/main/java/edu/umd/cs/findbugs/DetectorFactory.java
@@ -73,10 +73,6 @@ public class DetectorFactory {
 
     private String detailHTML;
 
-    private int priorityAdjustment;
-
-    private boolean enabledButNonReporting;
-
     private boolean hidden;
 
     static class ReflectionDetectorCreator {
@@ -172,7 +168,6 @@ public class DetectorFactory {
         this.speed = speed;
         this.reports = reports;
         this.requireJRE = requireJRE;
-        this.priorityAdjustment = 0;
         this.hidden = false;
     }
 
@@ -296,30 +291,13 @@ public class DetectorFactory {
         return defEnabled;
     }
 
-    /**
-     * Set the priority adjustment for the detector produced by this factory.
-     *
-     * @param priorityAdjustment
-     *            the priority adjustment
-     */
-    public void setPriorityAdjustment(int priorityAdjustment) {
-        this.priorityAdjustment = priorityAdjustment;
-    }
-
+    @Deprecated(forRemoval = true)
     public void setEnabledButNonReporting(boolean notReporting) {
-        this.enabledButNonReporting = notReporting;
     }
 
-    /**
-     * Get the priority adjustment for the detector produced by this factory.
-     *
-     * @return the priority adjustment
-     */
+    @Deprecated(forRemoval = true)
     public int getPriorityAdjustment() {
-        if (enabledButNonReporting) {
-            return 100;
-        }
-        return priorityAdjustment;
+        return 0;
     }
 
     /**