+++ spotbugs/src/main/java/edu/umd/cs/findbugs/config/UserPreferences.java
@@ -42,8 +42,10 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Properties;
 import java.util.Set;
+import java.util.StringTokenizer;
 import java.util.TreeMap;
 import java.util.TreeSet;
+import java.util.stream.Collectors;
 
 import javax.annotation.WillClose;
 
@@ -110,6 +112,8 @@ public class UserPreferences implements Cloneable {
 
     private static final String KEY_MERGE_SIMILAR_WARNINGS = "mergeSimilarWarnings";
 
+    private static final String KEY_ADJUST_PRIORITY = "adjust_priority";
+
     // Fields
 
     private LinkedList<String> recentProjectsList;
@@ -132,6 +136,8 @@ public class UserPreferences implements Cloneable {
 
     private boolean mergeSimilarWarnings;
 
+    private Map<String, String> adjustPriority;
+
     private UserPreferences() {
         filterSettings = ProjectFilterSettings.createDefault();
         recentProjectsList = new LinkedList<>();
@@ -142,6 +148,7 @@ public class UserPreferences implements Cloneable {
         excludeFilterFiles = new TreeMap<>();
         excludeBugsFiles = new TreeMap<>();
         customPlugins = new TreeMap<>();
+        adjustPriority = new TreeMap<>();
     }
 
     /**
@@ -255,6 +262,33 @@ public class UserPreferences implements Cloneable {
         excludeBugsFiles = readProperties(props, KEY_EXCLUDE_BUGS);
         customPlugins = readProperties(props, KEY_PLUGIN);
         mergeSimilarWarnings = Boolean.parseBoolean(props.getProperty(KEY_MERGE_SIMILAR_WARNINGS));
+        if (props.get(KEY_ADJUST_PRIORITY) != null) {
+            adjustPriority.putAll(parseAdjustPriorities(props.getProperty(KEY_ADJUST_PRIORITY)));
+        }
+    }
+
+    /**
+     * Parses the -adjustPriority parameter into a map: detector / bug pattern -> priority change
+     *
+     * @param str
+     *            the string to parse
+     * @return the parsed map
+     */
+    public static Map<String, String> parseAdjustPriorities(String str) {
+        Map<String, String> retVal = new HashMap<>();
+        StringTokenizer tokenizer = new StringTokenizer(str, ",");
+        while (tokenizer.hasMoreElements()) {
+            String token = tokenizer.nextToken();
+            int equalPos = token.indexOf('=');
+            if (equalPos > 0) {
+                String key = token.substring(0, equalPos).trim();
+                String value = token.substring(equalPos + 1).trim();
+                if (!key.isEmpty() && !value.isEmpty()) {
+                    retVal.put(key, value);
+                }
+            }
+        }
+        return retVal;
     }
 
     /**
@@ -308,7 +342,10 @@ public class UserPreferences implements Cloneable {
         writeProperties(props, KEY_EXCLUDE_FILTER, excludeFilterFiles);
         writeProperties(props, KEY_EXCLUDE_BUGS, excludeBugsFiles);
         writeProperties(props, KEY_PLUGIN, customPlugins);
-
+        props.put(KEY_ADJUST_PRIORITY, adjustPriority.entrySet()
+                .stream()
+                .map(entry -> entry.getKey() + "=" + entry.getValue())
+                .collect(Collectors.joining(",")));
         try (OutputStream prefStream = new BufferedOutputStream(out)) {
             props.store(prefStream, "SpotBugs User Preferences");
             prefStream.flush();
@@ -482,14 +519,16 @@ public class UserPreferences implements Cloneable {
                 && effort.equals(other.effort) && includeFilterFiles.equals(other.includeFilterFiles)
                 && excludeFilterFiles.equals(other.excludeFilterFiles) && excludeBugsFiles.equals(other.excludeBugsFiles)
                 && customPlugins.equals(other.customPlugins)
-                && mergeSimilarWarnings == other.mergeSimilarWarnings;
+                && mergeSimilarWarnings == other.mergeSimilarWarnings
+                && adjustPriority.equals(other.adjustPriority);
     }
 
     @Override
     public int hashCode() {
         return recentProjectsList.hashCode() + detectorEnablementMap.hashCode() + filterSettings.hashCode() + effort.hashCode()
                 + includeFilterFiles.hashCode() + excludeFilterFiles.hashCode() + (runAtFullBuild ? 1 : 0)
-                + excludeBugsFiles.hashCode() + customPlugins.hashCode() + (mergeSimilarWarnings ? 1 : 0);
+                + excludeBugsFiles.hashCode() + customPlugins.hashCode() + (mergeSimilarWarnings ? 1 : 0)
+                + adjustPriority.hashCode();
     }
 
     @Override
@@ -504,6 +543,7 @@ public class UserPreferences implements Cloneable {
             dup.excludeFilterFiles = new TreeMap<>(excludeFilterFiles);
             dup.excludeBugsFiles = new TreeMap<>(excludeBugsFiles);
             dup.customPlugins = new TreeMap<>(customPlugins);
+            dup.adjustPriority = new TreeMap<>(adjustPriority);
             return dup;
         } catch (CloneNotSupportedException e) {
             throw new AssertionError(e);
@@ -771,4 +811,20 @@ public class UserPreferences implements Cloneable {
         // no match, return the original path
         return maybeRelativePath;
     }
+
+    /**
+     * @return Returns the adjustPriority.
+     */
+    public Map<String, String> getAdjustPriority() {
+        return adjustPriority;
+    }
+
+    /**
+     * @param adjustPriority
+     *            The adjustPriority to set.
+     */
+    public void setAdjustPriority(Map<String, String> adjustPriority) {
+        this.adjustPriority = new TreeMap<>(adjustPriority);
+    }
+
 }