+++ spotbugs/src/main/java/edu/umd/cs/findbugs/io/IO.java
@@ -192,6 +192,27 @@ public class IO {
     }
 
     /**
+     * Close given InputStream, ignoring any resulting exception.
+     *
+     * @param inputStream
+     *            the InputStream to close; may be null (in which case nothing
+     *            happens)
+     * @deprecated This method is deprecated and will be removed in or after October 2026. Use try-with resources when creating the stream instead.
+     */
+    @Deprecated(forRemoval = true)
+    public static void close(@CheckForNull InputStream inputStream) {
+        if (inputStream == null) {
+            return;
+        }
+
+        try {
+            inputStream.close();
+        } catch (IOException e) {
+            // Ignore
+        }
+    }
+
+    /**
      * Close given OutputStream, ignoring any resulting exception.
      *
      * @param outputStream